now generating smallLookupA and smallLookupB correctly. when generating, it makes sense to compose using an entropy measure -- shorter codes correspond to more common sounds, longer codes to less common sounds

+++ trunk/cpp/mp3glib/src/Granule.h
@@ -15,8 +15,8 @@ public:
 	bool isShort() const;
 	bool isFirst() const;
 
-	static byte smallLookupA[81][2];
-	static byte smallLookupB[81][2];
+	static short smallLookupA[81][2];
+	static short smallLookupB[81][2];
 
 private:
 	Frame* frame;
